src/requirements.txt
# Core libraries
pandas==2.0.0  # Ensures you get a version 2.x or later
numpy==1.23.5
PyYAML==6.0

# Scikit-learn and related dependencies
scikit-learn==1.2.2
xgboost==1.7.6

# NLP and embedding transformer
sentence-transformers==2.2.2

# Miscellaneous
joblib==1.3.0